gitignore e BigDecimal

+++ src/Models/Entities/Despesa.java
@@ -2,13 +2,14 @@ package Models.Entities;
 
 import Models.Enums.Status;
 import Models.Enums.Metodo;
+import java.math.BigDecimal;
 import java.util.Date;
 
 public class Despesa extends Transacao
 {
 	protected String beneficiario;
 
-	public Despesa(int id, Date data, double valor, Metodo metodo, String descricao, int categoria, int subcategoria, Status status, String beneficiario)
+	public Despesa(int id, Date data, BigDecimal valor, Metodo metodo, String descricao, int categoria, int subcategoria, Status status, String beneficiario)
 	{
 		super(id, data, valor, metodo, descricao, categoria, subcategoria, status);
 		this.beneficiario = beneficiario;
@@ -16,7 +17,7 @@ public class Despesa extends Transacao
 
 
 	@Override
- 	public void update(Date data, double valor, Metodo metodo, String descricao, int categoria, int subcategoria, Status status, String interessado)
+ 	public void update(Date data, BigDecimal valor, Metodo metodo, String descricao, int categoria, int subcategoria, Status status, String interessado)
 	{
 		super.update(data, valor, metodo, descricao, categoria, subcategoria, status, null);
 		this.beneficiario = interessado;
