+++ src/Models/Entities/Receita.java
@@ -2,20 +2,21 @@ package Models.Entities;
 
 import Models.Enums.Metodo;
 import Models.Enums.Status;
+import java.math.BigDecimal;
 import java.util.Date;
 
 public class Receita extends Transacao
 {
 	protected String pagador;
 	
-	public Receita(int id, Date data, double valor, Metodo metodo, String descricao, int categoria, int subcategoria, Status status, String pagador)
+	public Receita(int id, Date data, BigDecimal valor, Metodo metodo, String descricao, int categoria, int subcategoria, Status status, String pagador)
 	{
 		super(id, data, valor, metodo, descricao, categoria, subcategoria, status);
 		this.pagador = pagador;
 	}
 
 	@Override
-	public void update(Date data, double valor, Metodo metodo, String descricao, int categoria, int subcategoria, Status status, String interessado)
+	public void update(Date data, BigDecimal valor, Metodo metodo, String descricao, int categoria, int subcategoria, Status status, String interessado)
 	{
 		super.update(data, valor, metodo, descricao, categoria, subcategoria, status, null);
 		this.pagador = interessado;
